+++ tests/System.Text.JsonLab.Tests/JsonWriterTests.cs
@@ -17,7 +17,7 @@ namespace System.Text.JsonLab.Tests
         public void WriteJsonUtf8()
         {
             var formatter = new ArrayFormatter(1024, SymbolTable.InvariantUtf8);
-            JsonWriter<ArrayFormatter> json = JsonWriter.Create(formatter, true, prettyPrint: false);
+            JsonWriterUtf8<ArrayFormatter> json = JsonWriter.CreateUtf8(formatter, prettyPrint: false);
             Write(ref json);
 
             var formatted = formatter.Formatted;
@@ -25,7 +25,7 @@ namespace System.Text.JsonLab.Tests
             Assert.Equal(expected, str.Replace(" ", ""));
 
             formatter.Clear();
-            json = JsonWriter.Create(formatter, true, prettyPrint: true);
+            json = JsonWriter.CreateUtf8(formatter, prettyPrint: true);
             Write(ref json);
 
             formatted = formatter.Formatted;
@@ -37,7 +37,7 @@ namespace System.Text.JsonLab.Tests
         public void WriteJsonUtf16()
         {
             var formatter = new ArrayFormatter(1024, SymbolTable.InvariantUtf16);
-            JsonWriter<ArrayFormatter> json = JsonWriter.Create(formatter, false, prettyPrint: false);
+            JsonWriterUtf16<ArrayFormatter> json = JsonWriter.CreateUtf16(formatter, prettyPrint: false);
             Write(ref json);
 
             var formatted = formatter.Formatted;
@@ -45,7 +45,7 @@ namespace System.Text.JsonLab.Tests
             Assert.Equal(expected, str.Replace(" ", ""));
 
             formatter.Clear();
-            json = JsonWriter.Create(formatter, false, prettyPrint: true);
+            json = JsonWriter.CreateUtf16(formatter, prettyPrint: true);
             Write(ref json);
 
             formatted = formatter.Formatted;
@@ -53,8 +53,32 @@ namespace System.Text.JsonLab.Tests
             Assert.Equal(expected, str.Replace("\r\n", "").Replace("\n", "").Replace(" ", ""));
         }
 
-        static string expected = "{\"age\":30,\"first\":\"John\",\"last\":\"Smith\",\"phoneNumbers\":[\"425-000-1212\",\"425-000-1213\",null],\"address\":{\"street\":\"1MicrosoftWay\",\"city\":\"Redmond\",\"zip\":98052},\"values\":[425121,-425122,425123]}";
-        static void Write(ref JsonWriter<ArrayFormatter> json)
+        static readonly string expected = "{\"age\":30,\"first\":\"John\",\"last\":\"Smith\",\"phoneNumbers\":[\"425-000-1212\",\"425-000-1213\",null],\"address\":{\"street\":\"1MicrosoftWay\",\"city\":\"Redmond\",\"zip\":98052},\"values\":[425121,-425122,425123]}";
+        static void Write(ref JsonWriterUtf8<ArrayFormatter> json)
+        {
+            json.WriteObjectStart();
+            json.WriteAttribute("age", 30);
+            json.WriteAttribute("first", "John");
+            json.WriteAttribute("last", "Smith");
+            json.WriteArrayStart("phoneNumbers");
+            json.WriteValue("425-000-1212");
+            json.WriteValue("425-000-1213");
+            json.WriteNull();
+            json.WriteArrayEnd();
+            json.WriteObjectStart("address");
+            json.WriteAttribute("street", "1 Microsoft Way");
+            json.WriteAttribute("city", "Redmond");
+            json.WriteAttribute("zip", 98052);
+            json.WriteObjectEnd();
+            json.WriteArrayStart("values");
+            json.WriteValue(425121);
+            json.WriteValue(-425122);
+            json.WriteValue(425123);
+            json.WriteArrayEnd();
+            json.WriteObjectEnd();
+            json.Flush();
+        }
+        static void Write(ref JsonWriterUtf16<ArrayFormatter> json)
         {
             json.WriteObjectStart();
             json.WriteAttribute("age", 30);
@@ -87,7 +111,7 @@ namespace System.Text.JsonLab.Tests
             string expectedStr = GetHelloWorldExpectedString(prettyPrint, isUtf8: false);
 
             var output = new ArrayFormatter(1024, SymbolTable.InvariantUtf16);
-            JsonWriter<ArrayFormatter> jsonUtf16 = JsonWriter.Create(output, false, prettyPrint);
+            JsonWriterUtf16<ArrayFormatter> jsonUtf16 = JsonWriter.CreateUtf16(output, prettyPrint);
 
             jsonUtf16.WriteObjectStart();
             jsonUtf16.WriteAttribute("message", "Hello, World!");
@@ -108,7 +132,7 @@ namespace System.Text.JsonLab.Tests
             string expectedStr = GetHelloWorldExpectedString(prettyPrint, isUtf8: true);
 
             var output = new ArrayFormatter(1024, SymbolTable.InvariantUtf8);
-            JsonWriter<ArrayFormatter> jsonUtf8 = JsonWriter.Create(output, true, prettyPrint);
+            JsonWriterUtf8<ArrayFormatter> jsonUtf8 = JsonWriter.CreateUtf8(output, prettyPrint);
 
             jsonUtf8.WriteObjectStart();
             jsonUtf8.WriteAttribute("message", "Hello, World!");
@@ -131,7 +155,7 @@ namespace System.Text.JsonLab.Tests
             string expectedStr = GetExpectedString(prettyPrint, isUtf8: false, data);
 
             var output = new ArrayFormatter(1024, SymbolTable.InvariantUtf16);
-            JsonWriter<ArrayFormatter> jsonUtf16 = JsonWriter.Create(output, false, prettyPrint);
+            JsonWriterUtf16<ArrayFormatter> jsonUtf16 = JsonWriter.CreateUtf16(output, prettyPrint);
 
             jsonUtf16.WriteObjectStart();
             jsonUtf16.WriteAttribute("age", 42);
@@ -175,7 +199,7 @@ namespace System.Text.JsonLab.Tests
             string expectedStr = GetExpectedString(prettyPrint, isUtf8: true, data);
 
             var output = new ArrayFormatter(1024, SymbolTable.InvariantUtf8);
-            JsonWriter<ArrayFormatter> jsonUtf8 = JsonWriter.Create(output, true, prettyPrint);
+            JsonWriterUtf8<ArrayFormatter> jsonUtf8 = JsonWriter.CreateUtf8(output, prettyPrint);
 
             jsonUtf8.WriteObjectStart();
             jsonUtf8.WriteAttribute("age", 42);
